v0.1.3 Made reboot optional

+++ scripts/fedora.sh
@@ -252,17 +252,20 @@ echo $'\e[1;32m'Процесс завершен, вот обзор вашей с
 echo $'\e[1;32m'プロセスが完了しました、 ここにあなたのシステムのレビューがあります。$'\e[0m'
 neofetch
 #asking for a reboot
-echo $'\e[1;31m'You should reboot to make sure the process is completed, do you want to reboot now?$'\e[0m'
-echo $'\e[1;31m'Debe reiniciar para asegurarse de que el proceso se haya completado, ¿desea reiniciar ahora?$'\e[0m'
-echo $'\e[1;31m'Вы должны перезагрузиться, чтобы убедиться, что процесс завершен, вы хотите перезагрузиться сейчас?$'\e[0m'
-echo $'\e[1;31m'プロセスが完了したことを確認するために再起動する必要があります。$'\e[0m'
-read reboot
-if [ $reboot == y ]
+if [ $nvidia == y ]
 then
-    sudo reboot
-else
-    echo $'\e[1;31m'The system will not be rebooted. The script has been concluded.$'\e[0m'
-    echo $'\e[1;31m'El sistema no se reiniciará. El script ha concluido.$'\e[0m'
-    echo $'\e[1;31m'Система не будет перезагружена. Сценарий завершен.$'\e[0m'
-    echo $'\e[1;31m'システムは再起動されません。スクリプトは終了しました。$'\e[0m'
+    echo $'\e[1;31m'You should reboot to make sure the process is completed, do you want to reboot now?$'\e[0m'
+    echo $'\e[1;31m'Debe reiniciar para asegurarse de que el proceso se haya completado, ¿desea reiniciar ahora?$'\e[0m'
+    echo $'\e[1;31m'Вы должны перезагрузиться, чтобы убедиться, что процесс завершен, вы хотите перезагрузиться сейчас?$'\e[0m'
+    echo $'\e[1;31m'プロセスが完了したことを確認するために再起動する必要があります。$'\e[0m'
+    read reboot
+    if [ $reboot == y ]
+    then
+        sudo reboot
+    else
+        echo $'\e[1;31m'The system will not be rebooted. The script has been concluded.$'\e[0m'
+        echo $'\e[1;31m'El sistema no se reiniciará. El script ha concluido.$'\e[0m'
+        echo $'\e[1;31m'Система не будет перезагружена. Сценарий завершен.$'\e[0m'
+        echo $'\e[1;31m'システムは再起動されません。スクリプトは終了しました。$'\e[0m'
+    fi
 fi
